requirements.txt
piper-tts>=1.3.0
numpy

